% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{confusion_plot}
\alias{confusion_plot}
\title{Confusion matrix plot}
\usage{
confusion_plot(d, fam, avgError, binom_labels = c(y0 =
  "observed class: 0", y1 = "observed class: 1", yhat0 =
  "predicted class: 0", yhat1 = "predicted class: 1"),
  multinom_labels = NULL)
}
\arguments{
\item{d}{Confusion matrix.}

\item{fam}{Model family.}

\item{avgError}{Average test error.}

\item{binom_labels}{Labels for the confusion matrix in the case of binomial
Classification.}

\item{multinom_labels}{Labels for the confusion matrix in the case of
multinomial Classification. Set to the column names of \code{d} if
no value is provided.}
}
\description{
Plots the confusion matrix.
}
