% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplit.R
\name{dataSplit}
\alias{dataSplit}
\title{Data split}
\usage{
dataSplit(y, train_size, valid_size, test_size, reps, balanced,
  fnrc = TRUE)
}
\arguments{
\item{y}{A numeric or factor response vector of size \code{N}, where \code{N} is
the number of observations (spectra).}

\item{train_size}{Proportion of observations to be assigned to the training subset.}

\item{valid_size}{Proportion of observations to be assigned to the validating subset.}

\item{test_size}{Proportion of observations to be assigned to the testing subset.}

\item{reps}{Number of randomisations of the training/validating/testing subsets to
produce.}

\item{balanced}{Whether the dataset should be balanced (TRUE) or not (FALSE).
See \code{\link{fdaML_train}} for more details.}

\item{fnrc}{Force Nonempty Response Class. Ensures that each replication has
samples from each possible level of the response.}
}
\value{
A list with the indices (running from \code{1} to \code{length(y)}) of
training, validation and testing observations. Each of these 3 lists is a list
of \code{reps} elements, each element being a vector of the length
\code{train_size}/\code{valid_size}/\code{test_size}, accordingly.
}
\description{
Split dataset into training, vaidation and testing subsets.
}
