% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaEstimation.R
\name{fdaEstimation}
\alias{fdaEstimation}
\title{Estimate parameters}
\usage{
fdaEstimation(XX, yy, optionsEst = NULL)
}
\arguments{
\item{XX}{A numeric predictor matrix.}

\item{yy}{A numeric response vector.}

\item{optionsEst}{A list with parameters for estimation.}
}
\description{
Produce parameter estimates for the statistical model.
}
