% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaML_predict.R
\name{fdaML_predict}
\alias{fdaML_predict}
\title{Predict from machine learning model}
\usage{
fdaML_predict(obj, new_x, new_z = NULL, new_y, verbose = TRUE)
}
\arguments{
\item{obj}{An object of class \code{fdaModel} generated by \code{\link{fdaML_train}}.}

\item{new_x}{A new functional predictor matrix.}

\item{new_z}{A new non-functional predictor matrix.}

\item{new_y}{A new response vector. If this is provided, the quality of predictions
are assessed.}

\item{verbose}{Whether to print a progress bar (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
An object of class \code{"fdaModelPred"}, which is a list
containing the results of the model.
}
\description{
Predict using machine learning model.
}
\details{

}
\examples{

}
\seealso{
\code{\link{fdaML_train}}
}
