% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaML_train.R
\name{fdaML_train}
\alias{fdaML_train}
\title{Train a Machine Learning model}
\usage{
fdaML_train(X, y, Z = NULL, task, model = NULL, reduction,
  intercept = TRUE, smooth_w = NULL, balanced = FALSE, reps = 100,
  Q_vec = NULL, Q_len = NULL, Q_opt = NULL, tau_Q_opt = 0,
  lam_cv_type = "n", lam_vec = NULL, split_size = c(0.5, 0.25, 0.25),
  estimation_w = NULL, bspline_dim = ncol(X), t_range = 350:2500,
  verbose = TRUE, ll = NULL)
}
\arguments{
\item{X}{A numeric functional predictor matrix of size \code{N*P}, where \code{N} is
the number of observations (spectra) and \code{P} is the number of points
(wavelengths) at which spectra are measured.}

\item{y}{A numeric or factor response vector of size \code{N}, where \code{N} is the number of
observations (spectra).}

\item{Z}{A numeric non-functional predictor matrix of size \code{N*S}, where \code{N}
is the number of observations (spectra) and \code{S} is the number of non-functional predictors
after bining/one-hot-encoding has taken place.}

\item{task}{Regression (\code{"regr"}) for continuous response problems or Classification
(\code{"clas"}) for categorical response problems.}

\item{model}{Currently Linear Model (\code{"lm"}) for Regression or Generalised Linear Model
(\code{"glm"}) for Classification.}

\item{reduction}{Partial Least Squares (\code{"pls"}), Principal Component Analysis
(\code{"pca"}), or no further dimension reduction (\code{"n"}).}

\item{intercept}{Whether to include a model intercept (\code{TRUE}) or not (\code{FALSE}).}

\item{smooth_w}{A numeric vector of length equal to \code{length(t_tange)} with weights for
smoothing spectra.}

\item{balanced}{Whether the dataset should be balanced (\code{TRUE}) or not (\code{FALSE}).
If \code{TRUE}, observations are discarded so that the number of observations for each
level of the response variable is approximately the same. It applies to both Classification
and Regression with integer-valued response.}

\item{reps}{Number of randomisations of the training/validating/testing subsets to average
over in cross-validation.}

\item{Q_vec}{Vector of numbers of PCA/PLS components to be tried in cross-validation.
The dedault is a vector of evenly spaced values (approximately, due to rounding) between
\code{2} and \code{min(80, N_train-1)}, where \code{N_train} is the number of observations
in the training subset.}

\item{Q_len}{Length of \code{Q_vec} when the later is not supplied. The default is 30.}

\item{Q_opt}{Optimal number of PCA/PLS components. If this is supplied, cross-validation for
the number of PCA/PLS components is bypassed.}

\item{tau_Q_opt}{Threshold for choosing the optimal parameters.
If \code{tau_Q_opt=0}, then \code{Q} is the value that minimises RMSD (for Regression)
or maximises AUC (for Classification).
If \code{tau_Q_opt>0}, then \code{Q} is the smallest value which gives a RMSD/AUC within
a margin \code{tau_Q_opt} of the optimal RMSD/AUC.}

\item{lam_cv_type}{Cross-validation strategy to be used when choosing the penalty
parameter \code{lambda}: ordinary cross-validation (\code{"ocv"}), generalised
cross-validation (\code{"gcv"}) or no penalisation (\code{"n"}).}

\item{lam_vec}{Vector of penalty parameters to be tried in cross-validation. The
default is a set of 10 values between 0.001 and 20 on an exponential scale.}

\item{split_size}{Either a vector of length 3 specifying the proportion of observations to be
assigned to the training, validation and testing subsets, in this order; or a scalar
specifying the proportion of observations to be assigned to the training subset, in which
case the validation and testing subsets are assigned a proportion \code{(1-split_size)/2}
of observations each.}

\item{estimation_w}{A numeric vector of length equal to \code{length(y)} with weights for
coeficient function estimation.}

\item{bspline_dim}{The dimension of the cubic B-spline functional representation system.}

\item{t_range}{A numeric vector giving the wavelenghts at which spectra were measured.}

\item{verbose}{Whether to print a progress bar (\code{TRUE}) or not (\code{FALSE}).}

\item{ll}{A list whose named elements are the parameters for this function.Provide either
the function parameters as usual, or this list, but not both.}
}
\value{
An object of class \code{fdaModel}, which is a list containing the trained model.
}
\description{
Train a machine learning model.
}
\details{

}
\examples{

}
\references{
P.M. Esperança, Thomas S. Churcher (2019). "Machine learning based
epidemiological vector control monitoring using functional data analysis
techniques for near-infrared spectral data".
arXiv.

P.T. Reiss, R.T. Ogden (2007). "Functional Principal Component Regression
and Functional Partial Least Squares".
Journal of the American Statistical Association, 102(479), 984-996
}
\seealso{
\code{\link{fdaML_predict}}
}
