% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPlot.R
\name{fdaPlot}
\alias{fdaPlot}
\title{Diagnostics and performance plots}
\usage{
fdaPlot(obj, hist_range = NULL, roc_smoothing = NULL,
  multinom_labels = NULL)
}
\arguments{
\item{obj}{An object of class \code{fdaModel} from \code{\link{fdaML_train}}
or class \code{fdaModelPred} from \code{\link{fdaML_predict}}.}

\item{hist_range}{A range for the histogram plot in binomial classification.}

\item{roc_smoothing}{A value for the parameter \code{spar} in \code{\link{smooth.spline}}.}

\item{multinom_labels}{Labels for the multinomial classes.}
}
\description{
Displays visual diagnostics information and performance measures
for objects of class \code{fdaModel} and \code{fdaModelPred}.
}
\seealso{
\code{\link{fdaML_train}} and \code{\link{fdaML_predict}}
}
