% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaPlotPred.R
\name{fdaPlotPred}
\alias{fdaPlotPred}
\title{Prediction results plot}
\usage{
fdaPlotPred(obj)
}
\arguments{
\item{obj}{An object of class \code{fdaModelPred} from \code{\link{fdaML_predict}}.}
}
\description{
Displays visual diagnostics information and performance measures
for objects of class \code{fdaModelPred}.
}
