% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GCV}
\alias{GCV}
\title{Generalised Cross-Validation for penalty parameter}
\usage{
GCV(XX, yy, penmat)
}
\arguments{
\item{XX}{Training predictor matrix.}

\item{yy}{Training response vector.}

\item{penmat}{Penalty matrix.}
}
\value{
The value of lambda which minimises the Generalised Cross-Validation criterion.
}
\description{
Generalised Cross-Validation method for optimising the penalty parameter.
}
\details{

}
\examples{

}
\references{

}
\seealso{

}
