% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_auc}
\alias{GET_auc}
\title{Compute AUC}
\usage{
GET_auc(y_pred, y_true, fam)
}
\arguments{
\item{y_pred}{Predicted response.}

\item{y_true}{True response.}

\item{fam}{Model family.}
}
\description{
Compute the Area Under the ROC Curve.
}
