% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_IdxQ_opt}
\alias{GET_IdxQ_opt}
\title{Optimal cross-validation parameters' index}
\usage{
GET_IdxQ_opt(perf, model, threshold)
}
\arguments{
\item{perf}{Performance matrix.}

\item{model}{Model type.}

\item{threshold}{Threshold for choosing the optimal parameters.
If \code{threshold=0}, then \code{Q} is the value that minimises RMSD (for Regression)
or maximises AUC (for Classification).
If \code{threshold>0}, then \code{Q} is the smallest value which gives a RMSD/AUC within
a margin \code{threshold} of the optimal RMSD/AUC.}
}
\description{
Compute index of best performing cross-validation parameters.
}
