% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_lambda}
\alias{GET_lambda}
\title{Compute penalty parameter}
\usage{
GET_lambda(XBD_trainQ, XBD_validQ, y_train, y_valid, optionEst)
}
\arguments{
\item{XBD_trainQ}{Training predictor matrix.}

\item{XBD_validQ}{Validation predictor matrix.}

\item{y_train}{Training response vector}

\item{y_valid}{Validation response matrix.}

\item{optionEst}{List of options for estimation.}
}
\description{
Compute penalty parameter by either Ordinary Cross-Validation
or GeneralisedCross-Validation.
}
