% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_optROCcutoff}
\alias{GET_optROCcutoff}
\title{Compute optimal cutoff for classification}
\usage{
GET_optROCcutoff(pred, perf, costs = NULL)
}
\arguments{
\item{pred}{An object of class \code{prediction} generated with \code{\link[ROCR]{prediction}}.}

\item{perf}{An object of class \code{performance} generated with \code{\link[ROCR]{performance}}.}

\item{costs}{A named list with elements \code{cost_fp} and \code{cost_fn} giving the weights
associated with false positives and false negatives, respectively.}
}
\description{
Compute the optimal cutoff for Classification
}
