% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_rmsd}
\alias{GET_rmsd}
\title{Compute RMSD}
\usage{
GET_rmsd(y_true, y_pred)
}
\arguments{
\item{y_true}{True response.}

\item{y_pred}{Predicted response.}
}
\value{
Root Mean Squared Deviation.
}
\description{
Compute the Root Mean Squared Deviation.
}
\details{

}
\examples{

}
\references{

}
\seealso{

}
