% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_roughness}
\alias{GET_roughness}
\title{Compute roughness of coefficient function}
\usage{
GET_roughness(x, std = F)
}
\arguments{
\item{x}{Coefficient function.}

\item{std}{Whether to standadise the coefficient function by its maximum
value (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Compute roughness of a coefficient function as given by the
integrated squared second derivative.
}
