% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_weights}
\alias{GET_weights}
\title{Compute weights}
\usage{
GET_weights(weighted, weights = NULL, yy, id, uniq)
}
\arguments{
\item{weighted}{Whether to use observation weights (\code{TRUE}) or not (\code{FALSE}).}

\item{weights}{Externally provided weights.}

\item{yy}{Response vector.}

\item{id}{Randomisation IDs.}

\item{uniq}{Unique levels in the response vector.}
}
\description{
Compute weights for each observation.
}
\details{
Compute weights invesely proportional to the number of observations
in each level of the response variable, if a vector of weights is not provided.
It applies to both Regression and Classification tasks.
}
