% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{LossFunction_GLM}
\alias{LossFunction_GLM}
\title{Penalised GLM objective function}
\usage{
LossFunction_GLM(pars, y_vec, XBD_mat, penmat, lam, l1, lv)
}
\arguments{
\item{pars}{Parameter vector (\code{alpha}, \code{omega}), with \code{alpha} scalar
and \code{omega} a \code{Q}-dimensional vector.}

\item{y_vec}{Response vector.}

\item{XBD_mat}{Functional predictor matrix after dimension reduction, plus an intercept
if appropriate.}

\item{penmat}{Penalty matrix.}

\item{lam}{Penalty parameter.}

\item{l1}{First column of functional predictor, equal to \code{1} if
\code{intercept=FALSE} and \code{2} if \code{intercept=TRUE}.}

\item{lv}{Last column of functional predictor, equal to\code{col(XBD_mat)}.}
}
\value{
The penalised GLM objective function evaluated at \code{pars}.
}
\description{
Objective function for the penalised GLM, to be optimised using stats::nlm().
}
\details{

}
\examples{

}
\references{

}
\seealso{

}
