% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlevcm-package.R
\docType{package}
\name{mlevcm-package}
\alias{mlevcm-package}
\alias{mlevcm}
\title{\code{mlevcm}: ML-based Epidemiological Vector Control Monitoring using
FDA techniques for NIRS data}
\description{
The \code{mlevcm} package provides functions to train and test machine
learning models capable of predicting characteristics of epidemiological
interest of disease vectors (e.g., mosquitoes) such as age, species and
infectiousness from near-infrared spectra.
}
\references{
P.M. Esperança, Thomas S. Churcher (2019). "Machine learning based
epidemiological vector control monitoring using functional data analysis
techniques for near-infrared spectral data".
arXiv.

P.T. Reiss, R.T. Ogden (2007). "Functional Principal Component Regression
and Functional Partial Least Squares".
Journal of the American Statistical Association, 102(479), 984-996
}
