% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{OCV}
\alias{OCV}
\title{Ordinary Cross-Validation for penalty parameter}
\usage{
OCV(XX, yy, newX, newY, optionsEst)
}
\arguments{
\item{XX}{Training predictor matrix.}

\item{yy}{Training response vector.}

\item{newX}{Validation predictor matrix.}

\item{newY}{Validation response matrix.}

\item{optionsEst}{List of options for estimation.}
}
\value{
The value of lambda which minimises the Ordinary Cross-Validation criterion,
that is, minimises the RMSD (for Regression) or maximises the AUC (for Classification).
}
\description{
Ordinary Cross-Validation method for optimising the penalty parameter.
}
\details{

}
\examples{

}
\references{

}
\seealso{

}
