% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplit.R
\name{dataSplit}
\alias{dataSplit}
\title{Data split}
\usage{
dataSplit(y, train_size, valid_size, test_size, reps, balanced,
  force_nonempty_response_class = TRUE)
}
\arguments{
\item{y}{A numeric or factor response vector of size \code{N}, where \code{N} is
the number of observations (spectra).}

\item{train_size}{The}

\item{valid_size}{}

\item{test_size}{}

\item{reps}{}

\item{balanced}{}

\item{force_nonempty_response_class}{}
}
\value{
A list with the indices (running from \code{1} to \code{length(y)}) of
training, validation and testing observations. Each of these 3 lists is a list
of \code{reps} elements, each element being a vector of the length
\code{train_size}/\code{valid_size}/\code{test_size}, accordingly.
}
\description{
Split dataset into training, vaidation and testing subsets.
}
