% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{GET_lambda}
\alias{GET_lambda}
\title{Compute penalty parameter}
\usage{
GET_lambda(XBD_trainQ, XBD_validQ, y_train, y_valid, optionsEst)
}
\arguments{
\item{XBD_trainQ}{Training predictor matrix.}

\item{XBD_validQ}{Validation predictor matrix.}

\item{y_train}{Training response vector}

\item{y_valid}{Validation response matrix.}

\item{optionsEst}{List of options for estimation.}
}
\value{
The value of the penalty parameter \code{lambda}.
}
\description{
Compute penalty parameter \code{lambda} by either ordinary
cross-validation (OCV) or generalised cross-validation (GCV).
}
\details{

}
\examples{

}
\references{

}
\seealso{

}
