% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{LossFunction_GLM}
\alias{LossFunction_GLM}
\title{Penalised GLM objective function}
\usage{
LossFunction_GLM(pars, y_vec, XBD_mat, S, intercept, penmat, lam)
}
\arguments{
\item{pars}{Parameter vector (\code{alpha}, \code{omega}), with \code{alpha} scalar
and \code{omega} a \code{Q}-dimensional vector.}

\item{y_vec}{Response vector.}

\item{XBD_mat}{Functional predictor matrix after dimension reduction.}

\item{S}{Number of non-functional predictors.}

\item{intercept}{Whether to include a model intercept (\code{TRUE}) or not (\code{FALSE}).}

\item{penmat}{Penalty matrix.}

\item{lam}{Penalty parameter.}
}
\description{
Objective function for the penalised GLM, to be optimised using stats::nlm().
}
