% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdaAuxFunctions.R
\name{OCV}
\alias{OCV}
\title{Ordinary Cross-Validation for penalty parameter}
\usage{
OCV(XX, yy, newX, newY, optionEst)
}
\arguments{
\item{XX}{Training predictor matrix.}

\item{yy}{Training response vector.}

\item{newX}{Validation predictor matrix.}

\item{newY}{Validation response matrix.}

\item{optionEst}{List of options for estimation.}
}
\description{
Ordinary Cross-Validation method for optimising the penalty parameter.
}
